Backend/requirements.txt
Flask
Flask-MySQLdb
Flask-CORS
mysql-connector-python 